app.backup/requirements.txt
streamlit==1.36.0
pandas==2.2.2
python-dateutil==2.9.0.post0
numpy==1.26.4
matplotlib==3.8.4
